kernel: thread: move k_thread_priority_get
Move to thread.c alongside all other thread calls.
