+++ kernel/sched.c
@@ -1025,20 +1025,6 @@ void z_sched_init(void)
 #endif /* CONFIG_SCHED_CPU_MASK_PIN_ONLY */
 }
 
-int z_impl_k_thread_priority_get(k_tid_t thread)
-{
-	return thread->base.prio;
-}
-
-#ifdef CONFIG_USERSPACE
-static inline int z_vrfy_k_thread_priority_get(k_tid_t thread)
-{
-	K_OOPS(K_SYSCALL_OBJ(thread, K_OBJ_THREAD));
-	return z_impl_k_thread_priority_get(thread);
-}
-#include <syscalls/k_thread_priority_get_mrsh.c>
-#endif /* CONFIG_USERSPACE */
-
 void z_impl_k_thread_priority_set(k_tid_t thread, int prio)
 {
 	/*
